feat(Simulator): Moved simulator to be inside the UV environment to be using the the path planner
